+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `DataItemDto.serializeResult` function
 
 ## 5.7.0 - 2021-09-27
 - Update dependencies